x86: nVMX: Dynamically calculate and check max VMCS field encoding index
Per Intel's SDM:

  IA32_VMX_VMCS_ENUM indicates to software the highest index value used
  in the encoding of any field supported by the processor:
    - Bits 9:1 contain the highest index value used for any VMCS encoding.
    - Bit 0 and bits 63:10 are reserved and are read as 0

KVM correctly emulates this behavior, in no small part due to the VMX
preemption timer being unconditionally emulated *and* having the highest
index of any field supported in vmcs12.  Given that the maximum control
field index is already above the VMX preemption timer (0x32 vs 0x2E),
odds are good that the max index supported in vmcs12 will change in the
not-too-distant future.

Unfortunately, the only unit test coverage for IA32_VMX_VMCS_ENUM is in
test_vmx_caps(), which simply checks that the max index is >= 0x2a, i.e.
won't catch any future breakage of KVM's IA32_VMX_VMCS_ENUM emulation,
especially if the max index depends on underlying hardware support.

Instead of playing whack-a-mole with a hardcoded max index test,
piggyback the exhaustive VMWRITE/VMREAD test and dynamically calculate
the max index based on which fields can be VMREAD.  Leave the existing
hardcoded check in place as it won't hurt anything and test_vmx_caps()
is a better location for checking the reserved bits of the MSR.

Cc: Nadav Amit <nadav.amit@gmail.com>
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>